api/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.117.1
h11==0.16.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
email-validator>=2.1
python-dotenv>=1.0
